data type change customerId by pavan

+++ A-Dream-BankManagementSystem-LoanApplicationAndVarificationService/src/main/java/com/dbms/loanapplicationandvarification/main/model/Customer.java
@@ -27,7 +27,7 @@ public class Customer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long customerId;
+    private int customerId;
 
     @NotNull(message = "You Can Not Take The Customer FirstName Is Null !!!")
 	@NotBlank(message = "Customer FirstName Can Not Be Blank")
